Add the command-line parser class
The class defines a first set of methods, two addOption overloads to
define a nullary or unary command-line option (each overload accepts
the name of the option and either a nullary or unary callback), and a
parse method, to run the parser on the provided command-line arguments.

The methods are documented, the class too but incompletely.
The method bodies are not implemented yet, they only throw an
exception.